- Warnings and a fix for some ARM32 issues.

+++ StereoKitC/asset_types/material.cpp
@@ -727,7 +727,7 @@ void material_check_tex_changes(material_t material) {
 	// Textures that progressively load or swap from fallbacks will change
 	// their dimensions. SK provides dimensions via the texname_i variable, so
 	// we're making sure that stays in sync here.
-	for (size_t i = 0; i < material->args.texture_count; i++) {
+	for (int32_t i = 0; i < material->args.texture_count; i++) {
 		shaderargs_tex_t *curr = &material->args.textures[i];
 		if (curr->meta_hash != curr->tex->meta_hash) {
 			tex_t physical_tex = curr->tex->fallback == nullptr
